moved constant and function files
