+++ packages/destination-actions/src/destinations/google-enhanced-conversions/uploadClickConversion/index.ts
@@ -2,7 +2,7 @@ import { ActionDefinition, IntegrationError } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
 import { CartItem } from './types'
-import { formatCustomVariables, verifyCurrency, hash } from './functions'
+import { formatCustomVariables, verifyCurrency, hash } from '../functions'
 
 const action: ActionDefinition<Settings, Payload> = {
   title: 'Upload Click Conversion',
